desktop/requirements.txt
firebase-admin==6.2.0
PyQt6==6.4.0
PyQt6-Qt6==6.4.0
PyQt6-sip==13.4.0
python-dotenv==0.21.0
requests==2.28.1
plyer==2.1.0
google-cloud-storage==2.10.0
pyaudio==0.2.11